Unit tests for subgraph query mapping
SPARQL queries are mapped to the identified wikidata subgraphs
based on several criteria (items, properties usage etc) to
estimate which subgraphs are accessed by these queries.
This patch contains tests that use a small set of queries
on a small wikidata subset to identify which subgraphs are
accessed by these queries.

Bug: T303831

+++ rdf-spark-tools/src/main/scala/org/wikidata/query/rdf/spark/transform/queries/subgraphsqueries/SubgraphQueryMappingLauncher.scala
@@ -11,7 +11,7 @@ case class SubgraphQueryMappingParams(
                                        subgraphPredicatesMatchInQuery: String = "", //out
                                        subgraphUrisMatchInQuery: String = "", //out
                                        queryMapping: String = "", //out
-                                       filteringLimit: Int = 99
+                                       filteringLimit: Double = 99
                                      )
 
 /**
